Optimized speed of weather_widget and weather_widget_mini by introducing common shared data between them
