+++ lib/widgets/weather_widget_mini.dart
@@ -1,16 +1,17 @@
-// weather_widget.dart
-import 'package:flutter/material.dart';
-import './services/weather_services.dart';
+// weather_widget_mini.dart
 
-class WeatherWidget extends StatefulWidget {
-  final VoidCallback? onClose;
-  const WeatherWidget({Key? key, this.onClose}) : super(key: key);
+import 'package:flutter/material.dart';
+import './services/weather_service.dart';
+import 'weather_widget.dart'; // Add this import
+
+class MiniWeatherWidget extends StatefulWidget {
+  const MiniWeatherWidget({Key? key}) : super(key: key);
 
   @override
-  _WeatherWidgetState createState() => _WeatherWidgetState();
+  _MiniWeatherWidgetState createState() => _MiniWeatherWidgetState();
 }
 
-class _WeatherWidgetState extends State<WeatherWidget> {
+class _MiniWeatherWidgetState extends State<MiniWeatherWidget> {
   final WeatherService _weatherService = WeatherService();
   Map<String, dynamic>? _weatherData;
   String? _error;
@@ -32,7 +33,7 @@ class _WeatherWidgetState extends State<WeatherWidget> {
 
   Future<void> _initialize() async {
     setState(() => _loading = true);
-    
+
     try {
       final locationData = await _weatherService.initializeLocation();
       if (locationData != null) {
@@ -50,218 +51,71 @@ class _WeatherWidgetState extends State<WeatherWidget> {
     }
   }
 
+  void _showWeatherDialog() {
+    showDialog(
+      context: context,
+      builder: (context) => WeatherWidget(
+        onClose: () => Navigator.of(context).pop(),
+      ),
+    );
+  }
+
   @override
   Widget build(BuildContext context) {
-    return Center(
-      child: Container(
-        constraints: const BoxConstraints(maxWidth: 400),
-        decoration: BoxDecoration(
-          color: Colors.black87,
-          borderRadius: BorderRadius.circular(15),
-          border: Border.all(color: Colors.white24, width: 1),
-        ),
-        child: Stack(
-          children: [
-            Padding(
-              padding: const EdgeInsets.all(16.0),
-              child: _buildContent(),
+    return GestureDetector(
+      onTap: _showWeatherDialog,
+      child: LayoutBuilder(
+        builder: (context, constraints) {
+          final iconSize = constraints.maxWidth * 0.15;
+          final tempSize = constraints.maxWidth * 0.12;
+
+          return Container(
+            decoration: BoxDecoration(
+              color: Colors.black87,
+              borderRadius: BorderRadius.circular(15),
             ),
-            if (widget.onClose != null)
-              Positioned(
-                right: 8,
-                top: 8,
-                child: IconButton(
-                  icon: const Icon(Icons.close, color: Colors.white70),
-                  onPressed: widget.onClose,
-                  constraints: const BoxConstraints(),
-                  padding: EdgeInsets.zero,
-                ),
-              ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildContent() {
-    if (_loading) {
-      return const SizedBox(
-        height: 120,
-        child: Center(child: CircularProgressIndicator(color: Colors.white)),
-      );
-    }
-
-    if (_error != null) {
-      return _buildErrorWidget();
-    }
-
-    if (_weatherData != null) {
-      return _buildWeatherInfo();
-    }
-
-    return const SizedBox(
-      height: 120,
-      child: Center(
-        child: Text(
-          'No weather data available',
-          style: TextStyle(color: Colors.white70),
-        ),
-      ),
-    );
-  }
-
-  Widget _buildWeatherInfo() {
-    final weatherDescription = _weatherData!['weather'][0]['description'];
-    final IconData weatherIcon = WeatherService.weatherIcons[weatherDescription] ?? 
-        Icons.question_mark_rounded;
-
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Expanded(
+            child: Padding(
+              padding: const EdgeInsets.all(12.0),
               child: Row(
+                mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: [
-                  const Icon(Icons.location_on_rounded, 
-                    color: Colors.white70, size: 20),
-                  const SizedBox(width: 8),
-                  Expanded(
-                    child: Text(
-                      _weatherData!['name'],
-                      style: const TextStyle(
-                        fontSize: 20,
-                        color: Colors.white,
-                        fontWeight: FontWeight.w500,
+                  if (_loading)
+                    const CircularProgressIndicator(color: Colors.white)
+                  else if (_error != null)
+                    Icon(
+                      Icons.error_outline,
+                      color: Colors.redAccent,
+                      size: iconSize.clamp(24.0, 32.0),
+                    )
+                  else if (_weatherData != null)
+                    Expanded(
+                      child: Row(
+                        children: [
+                          Icon(
+                            WeatherService.weatherIcons[_weatherData!['weather']
+                                    [0]['description']] ??
+                                Icons.wb_sunny_rounded,
+                            color: Colors.white,
+                            size: iconSize.clamp(24.0, 32.0),
+                          ),
+                          const SizedBox(width: 12),
+                          Text(
+                            '${_weatherData!['main']['temp'].round()}°C',
+                            style: TextStyle(
+                              fontSize: tempSize.clamp(20.0, 28.0),
+                              color: Colors.white,
+                              fontWeight: FontWeight.bold,
+                            ),
+                          ),
+                        ],
                       ),
-                      overflow: TextOverflow.ellipsis,
                     ),
-                  ),
                 ],
               ),
             ),
-            IconButton(
-              icon: const Icon(Icons.refresh_rounded, color: Colors.white70),
-              onPressed: _initialize,
-              padding: EdgeInsets.zero,
-              constraints: const BoxConstraints(),
-            ),
-          ],
-        ),
-        const SizedBox(height: 16),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Icon(weatherIcon, color: Colors.white, size: 40),
-            const SizedBox(width: 16),
-            Text(
-              '${_weatherData!['main']['temp'].round()}°C',
-              style: const TextStyle(
-                fontSize: 36,
-                color: Colors.white,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 8),
-        Text(
-          weatherDescription.toString().toUpperCase(),
-          style: const TextStyle(
-            color: Colors.white70,
-            fontSize: 14,
-            letterSpacing: 1.2,
-          ),
-          textAlign: TextAlign.center,
-        ),
-        const SizedBox(height: 16),
-        _buildWeatherDetails(),
-      ],
-    );
-  }
-
-  Widget _buildWeatherDetails() {
-    return Container(
-      padding: const EdgeInsets.symmetric(vertical: 12, horizontal: 8),
-      decoration: BoxDecoration(
-        color: Colors.white.withOpacity(0.1),
-        borderRadius: BorderRadius.circular(12),
+          );
+        },
       ),
-      child: IntrinsicHeight(
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: [
-            Expanded(
-              child: _buildDetailColumn(
-                Icons.water_drop_outlined,
-                'Humidity',
-                '${_weatherData!['main']['humidity']}%',
-              ),
-            ),
-            VerticalDivider(color: Colors.white.withOpacity(0.2), thickness: 1),
-            Expanded(
-              child: _buildDetailColumn(
-                Icons.air_rounded,
-                'Wind',
-                '${_weatherData!['wind']['speed']} m/s',
-              ),
-            ),
-            VerticalDivider(color: Colors.white.withOpacity(0.2), thickness: 1),
-            Expanded(
-              child: _buildDetailColumn(
-                Icons.thermostat_rounded,
-                'Feels Like',
-                '${_weatherData!['main']['feels_like'].round()}°C',
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-
-  Widget _buildDetailColumn(IconData icon, String label, String value) {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        Icon(icon, color: Colors.white70, size: 20),
-        const SizedBox(height: 4),
-        Text(
-          label,
-          style: const TextStyle(color: Colors.white70, fontSize: 12),
-        ),
-        const SizedBox(height: 2),
-        Text(
-          value,
-          style: const TextStyle(
-            color: Colors.white,
-            fontSize: 14,
-            fontWeight: FontWeight.w500,
-          ),
-        ),
-      ],
-    );
-  }
-
-  Widget _buildErrorWidget() {
-    return Column(
-      mainAxisSize: MainAxisSize.min,
-      children: [
-        const Icon(Icons.error_outline, color: Colors.redAccent, size: 32),
-        const SizedBox(height: 8),
-        Text(
-          _error ?? 'Error loading weather data',
-          style: const TextStyle(color: Colors.white70, fontSize: 14),
-        ),
-        const SizedBox(height: 12),
-        TextButton.icon(
-          onPressed: _initialize,
-          icon: const Icon(Icons.refresh, color: Colors.white),
-          label: const Text('Try Again', style: TextStyle(color: Colors.white)),
-        ),
-      ],
     );
   }
 }
